requirements.txt
flask
google.generativeai
flask_session
